Add apple IIGS line length compensation and change FBsize values for clarity

+++ src/defs.h
@@ -194,7 +194,7 @@ typedef struct {
 
 typedef struct {
    int clock;                // sample clock frequency (Hz)
-   int line_len;             // length of a line (in sample clocks)
+   double line_len;          // length of a line (in sample clocks)
    int clock_ppm;            // sample clock frequency (Hz)
    int lines_per_frame;
 } clk_info_t;
